Parser Module Added
* Parser Module created, will add majority of parser stuff in it
* Added loaded message for modules
